+++ crates/gen/src/types/com_interface.rs
@@ -19,7 +19,7 @@ impl ComInterface {
                 .chain(std::iter::once(&self.0))
                 .map(|def| def.methods())
                 .flatten()
-                .map(|method| method.signature(&[]).gen_win32_abi(gen));
+                .map(|method| gen_win32_abi(&method.signature(&[]), gen));
 
             let mut method_names = BTreeMap::<String, u32>::new();
